[API-Compat] Split compat.py into a folder of separate func impls
