+++ python/setup.py.in
@@ -1001,6 +1001,9 @@ packages=['paddle',
           'paddle.quantization',
           'paddle.quantization.imperative',
           'paddle.tensor',
+          'paddle.compat',
+          'paddle.compat.nn',
+          'paddle.compat.nn.functional',
           'paddle.onnx',
           'paddle.autograd',
           'paddle.cuda',
