+++ setup.py
@@ -2464,6 +2464,9 @@ def get_setup_parameters():
         'paddle.quantization',
         'paddle.quantization.imperative',
         'paddle.tensor',
+        'paddle.compat',
+        'paddle.compat.nn',
+        'paddle.compat.nn.functional',
         'paddle.onnx',
         'paddle.autograd',
         'paddle.cuda',
